# Курсовая/CursovoiProject/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API server for Veterinary Clinic management system.",
        "title": "Veterinary Clinic API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api",
    "paths": {
        "/appointments": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Бронирует визит к врачу",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "booking"
                ],
                "summary": "Создание записи на прием",
                "parameters": [
                    {
                        "description": "Данные записи",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateAppointmentRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    }
                }
            }
        },
        "/appointments/busy-slots": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Возвращает список времени, на которое уже есть запись у врача",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "booking"
                ],
                "summary": "Получение занятых слотов",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID врача",
                        "name": "doctor_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Дата (YYYY-MM-DD)",
                        "name": "date",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/appointments/client/{userId}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "booking"
                ],
                "summary": "История визитов клиента",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID пользователя",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.RecentAppointmentDTO"
                            }
                        }
                    }
                }
            }
        },
...
        "/auth/login": {
            "post": {
                "description": "Аутентификация и получение JWT токена",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Вход в систему",
                "parameters": [
                    {
                        "description": "Данные входа",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.LoginResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "Создает новый аккаунт клиента",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Регистрация пользователя",
                "parameters": [
                    {
                        "description": "Данные регистрации",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/book/init/{userId}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Получает списки питомцев, врачей и услуг",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "booking"
                ],
                "summary": "Инициализация формы записи",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID пользователя",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.BookingInitResponse"
                        }
                    }
                }
            }
        },
...
        "/dashboard/{userId}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Получает сводку статистики, ближайшую запись и список питомцев",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboard"
                ],
                "summary": "Данные дашборда",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID пользователя",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.DashboardResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.BookingInitResponse": {
            "type": "object",
            "properties": {
                "doctors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.DoctorDTO"
                    }
                },
                "pets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.PetBriefDTO"
                    }
                },
                "services": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ServiceDTO"
                    }
                }
            }
        },
...
        "dto.CreateAppointmentRequest": {
            "type": "object",
            "required": [
                "doctor_id",
                "pet_id",
                "scheduled_at"
            ],
            "properties": {
                "comment": {
                    "type": "string"
                },
                "doctor_id": {
                    "type": "integer"
                },
                "pet_id": {
                    "type": "integer"
                },
                "scheduled_at": {
                    "description": "\"2024-05-20T09:00:00\"",
                    "type": "string"
                },
                "service_id": {
                    "type": "integer"
                }
            }
        },
        "dto.DashboardResponse": {
            "type": "object",
            "properties": {
                "appointmentsCount": {
                    "type": "integer"
                },
                "nextAppointment": {
                    "$ref": "#/definitions/dto.NextAppointmentDTO"
                },
                "pets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.PetBriefDTO"
                    }
                },
                "petsCount": {
                    "type": "integer"
                },
                "recentAppointments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.RecentAppointmentDTO"
                    }
                },
                "visitsCount": {
                    "type": "integer"
                }
            }
        },
        "dto.DiagnosisDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "dto.DoctorDTO": {
            "type": "object",
            "properties": {
                "doctorId": {
                    "type": "integer"
                },
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "specialization": {
                    "type": "string"
                }
            }
        },
...
        "dto.LoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "dto.LoginResponse": {
            "type": "object",
            "properties": {
                "lastName": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "userId": {
                    "type": "integer"
                },
                "userName": {
                    "type": "string"
                }
            }
        },
        "dto.MedicalProtocolDTO": {
            "type": "object",
            "properties": {
                "diagnoses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.DiagnosisDTO"
                    }
                },
                "medications": {
                    "type": "string"
                },
                "treatment": {
                    "type": "string"
                },
                "weight": {
                    "type": "number"
                }
            }
        },
        "dto.NextAppointmentDTO": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "petName": {
                    "type": "string"
                },
                "time": {
                    "type": "string"
                }
            }
        },
...
        "dto.PetBriefDTO": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "breed": {
                    "type": "string"
                },
                "dob": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "weight": {
                    "type": "number"
                }
            }
        },
...
        "dto.RecentAppointmentDTO": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "doctorName": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "petName": {
                    "type": "string"
                },
                "protocol": {
                    "$ref": "#/definitions/dto.MedicalProtocolDTO"
                },
                "service": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "time": {
                    "type": "string"
                }
            }
        },
        "dto.RegisterRequest": {
            "type": "object",
            "required": [
                "email",
                "firstName",
                "lastName",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 8
                },
                "phone": {
                    "type": "string"
                }
            }
        },
        "dto.ServiceDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}